+++ bank-feeds/src/codat_bankfeeds/models/shared/bankfeedaccountmapping.py
@@ -10,12 +10,34 @@ from codat_bankfeeds.types import (
 )
 import pydantic
 from pydantic import model_serializer
+from typing import Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class BankFeedAccountMappingTypedDict(TypedDict):
     source_account_id: str
     r"""Unique ID for the source account"""
+    feed_start_date: NotRequired[str]
+    r"""In Codat's data model, dates and times are represented using the <a class=\"external\" href=\"https://en.wikipedia.org/wiki/ISO_8601\" target=\"_blank\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
+
+    ```
+    2020-10-08T22:40:50Z
+    2021-01-01T00:00:00
+    ```
+
+
+
+    When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
+
+    - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
+    - Unqualified local time: `2021-11-15T01:00:00`
+    - UTC time offsets: `2021-11-15T01:00:00-05:00`
+
+    > Time zones
+    >
+    > Not all dates from Codat will contain information about time zones.
+    > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
+    """
     target_account_id: NotRequired[Nullable[str]]
     r"""Unique ID for the target account"""
 
@@ -24,6 +46,30 @@ class BankFeedAccountMapping(BaseModel):
     source_account_id: Annotated[str, pydantic.Field(alias="sourceAccountId")]
     r"""Unique ID for the source account"""
 
+    feed_start_date: Annotated[Optional[str], pydantic.Field(alias="feedStartDate")] = (
+        None
+    )
+    r"""In Codat's data model, dates and times are represented using the <a class=\"external\" href=\"https://en.wikipedia.org/wiki/ISO_8601\" target=\"_blank\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
+
+    ```
+    2020-10-08T22:40:50Z
+    2021-01-01T00:00:00
+    ```
+
+
+
+    When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
+
+    - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
+    - Unqualified local time: `2021-11-15T01:00:00`
+    - UTC time offsets: `2021-11-15T01:00:00-05:00`
+
+    > Time zones
+    >
+    > Not all dates from Codat will contain information about time zones.
+    > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
+    """
+
     target_account_id: Annotated[
         OptionalNullable[str], pydantic.Field(alias="targetAccountId")
     ] = UNSET
@@ -31,7 +77,7 @@ class BankFeedAccountMapping(BaseModel):
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = ["targetAccountId"]
+        optional_fields = ["feedStartDate", "targetAccountId"]
         nullable_fields = ["targetAccountId"]
         null_default_fields = []
 
@@ -39,7 +85,7 @@ class BankFeedAccountMapping(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)