+++ bank-feeds/src/codat_bankfeeds/models/shared/dataconnectionerror.py
@@ -153,7 +153,7 @@ class DataConnectionError(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)