+++ bank-feeds/RELEASES.md
@@ -879,3 +879,13 @@ Based on:
 - [python v11.0.0] bank-feeds
 ### Releases
 - [PyPI v11.0.0] https://pypi.org/project/codat-bankfeeds/11.0.0 - bank-feeds
+
+## 2025-04-09 10:08:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.531.0 (2.568.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v11.1.0] bank-feeds
+### Releases
+- [PyPI v11.1.0] https://pypi.org/project/codat-bankfeeds/11.1.0 - bank-feeds